+++ install.sh
@@ -250,7 +250,7 @@ fi
 require_brew nvm
 
 # nvm
-require_nvm 4.4.7
+require_nvm stable
 
 # always pin versions (no surprises, consistent dev/build machines)
 npm config set save-exact true