+++ HISTORY.md
@@ -32,6 +32,7 @@
  * make gitshots optional (and only install `imagesnap` and `imagemagick` if this is on)
  * fix passwordless sudo for Sierra (optional)
  * prompt to overwrite /etc/hosts with someonewhocares.org hosts file (saved in ./configs/hosts in this project)
+ * fix default wallpaper in Sierra (Sierra 2)
 
 ## v4.3.0
  * vim installation fix