+++ HISTORY.md
@@ -33,6 +33,8 @@
  * fix passwordless sudo for Sierra (optional)
  * prompt to overwrite /etc/hosts with someonewhocares.org hosts file (saved in ./configs/hosts in this project)
  * fix default wallpaper in Sierra (Sierra 2)
+ * remove sudden motion disable (already disabled: https://github.com/mathiasbynens/dotfiles/pull/755/files)
+ * no longer setting hibernation mode to 0 (leaving default 3): https://github.com/mathiasbynens/dotfiles/commit/a3f91f67e07b6b31760b52320e0e890f93ff4e97#commitcomment-20715991
 
 ## v4.3.0
  * vim installation fix